prova_backend/projeto/requirements.txt
Django==2.2.16
djangorestframework==3.9.0